pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "greyhounds-ann-lab"
version = "0.1.0"
description = "Greyhound race ingestion, ANN training, and prediction tooling."
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
  "numpy>=1.26,<3",
  "pandas>=2.2,<3",
  "python-dotenv>=1.0,<2",
  "requests>=2.32,<3",
  "SQLAlchemy>=2.0,<3",
  "psycopg[binary]>=3.2,<4",
  "streamlit>=1.44,<2",
  "torch>=2.4,<3",
  "typer>=0.12,<1",
]

[project.scripts]
greyhounds = "greyhounds.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
